pyproject.toml
[project]
name = "quali_analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bitarray>=3.7.1",
    "hydra-core>=1.3.2",
    "ipdb>=0.13.13",
    "matplotlib>=3.10.6",
    "omegaconf>=2.3.0",
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "prettytable>=3.16.0",
    "sacrebleu>=2.5.1",
    "scikit-learn>=1.7.2",
    "scipy>=1.15.3",
    "streamlit>=0.1.0",
    "torch==2.7.0",
    "transformers==4.52.4",
    "wandb>=0.22.1",
]
